[9-1] Refactoring
