+++ gogobikeos/bootpack.h
@@ -105,8 +105,6 @@ void mysprintf(char *str, char *fmt, ...);
  * int.c
  */
 void init_pic(void);
-void inthandler21(int *esp);
-void inthandler2c(int *esp);
 #define PIC0_ICW1       0x0020
 #define PIC0_OCW2       0x0020
 #define PIC0_IMR        0x0021
@@ -119,3 +117,26 @@ void inthandler2c(int *esp);
 #define PIC1_ICW2       0x00a1
 #define PIC1_ICW3       0x00a1
 #define PIC1_ICW4       0x00a1
+
+/*
+ * keyboard.c
+ */
+void inthandler21(int *esp);
+void wait_KBC_sendready(void);
+void init_keyboard(void);
+extern struct FIFO8 keyfifo;
+#define PORT_KEYDAT             0x0060
+#define PORT_KEYCMD             0x0064
+
+/*
+ * mouse.c
+ */
+
+struct MOUSE_DEC {
+    unsigned char buf[3], phase;
+    int x, y, btn;
+};
+void inthandler2c(int *esp);
+void enable_mouse(struct MOUSE_DEC *mdec);
+int mouse_decode(struct MOUSE_DEC *mdec, unsigned char dat);
+extern struct FIFO8 mousefifo;
