+++ gogobikeos/int.c
@@ -26,32 +26,3 @@ void init_pic(void)
 
     return;
 }
-
-struct FIFO8 keyfifo;
-
-/*
- * PS/2 キーボードからの割り込み
- */
-void inthandler21(int *esp)
-{
-    unsigned char data;
-    io_out8(PIC0_OCW2, 0x61); // IRQ-01 受付完了を PIC に通知
-    data = io_in8(PORT_KEYDAT);
-    fifo8_put(&keyfifo, data);
-    return;
-}
-
-struct FIFO8 mousefifo;
-
-/*
- * PS/2 マウスからの割り込み
- */
-void inthandler2c(int *esp)
-{
-    unsigned char data;
-    io_out8(PIC1_OCW2, 0x64);   // IRQ-12 受付完了を PIC1 に通知
-    io_out8(PIC0_OCW2, 0x62);   // IRQ-02 受付完了を PIC0 に通知
-    data = io_in8(PORT_KEYDAT);
-    fifo8_put(&mousefifo, data);
-    return;
-}
